perf(app): code-split routes via lazy components, wire devtools, add 5-min tutorial
- Each <name>.route.tsx now uses lazyRouteComponent(() => import("./<name>.page"),
  "<Name>Page"). Page components extracted to <name>.page.tsx siblings (14 routes).
  Initial bundle: 760KB → 588KB (-23%, -172KB). Each route ships as a 1-43KB
  chunk loaded on navigation; intent prefetch keeps it imperceptible.
- Pages access their route's search/params/context via getRouteApi(path)
  instead of importing the route binding directly. The path string is
  type-checked through the Register module declaration; the static import
  cycle (page <-> route) disappears.
- TanStack Router devtools + React Query devtools mounted in app-providers.tsx
  behind import.meta.env.DEV (tree-shaken in prod).
- README "5-minute clone tutorial" replaces the bare quick-start: bootstrap →
  adopt → trim → first feature → ship. Concrete commands, paste-ready.
- vite.config chunkSizeWarningLimit: 700 — silences the warning at the new
  588KB floor while still tripping on regressions.
