+++ apps/app/package.json
@@ -34,6 +34,8 @@
   "devDependencies": {
     "@packages/typescript-config": "workspace:*",
     "@tailwindcss/vite": "^4.2.4",
+    "@tanstack/react-query-devtools": "^5.100.9",
+    "@tanstack/react-router-devtools": "^1.166.13",
     "@types/node": "^25.6.0",
     "@types/react": "^19.2.14",
     "@types/react-dom": "^19.2.3",
