+++ apps/app/src/features/auth/sign-up.route.tsx
@@ -1,29 +1,8 @@
-import { createRoute, Link } from "@tanstack/react-router";
+import { createRoute, lazyRouteComponent } from "@tanstack/react-router";
 import { guestLayout } from "../../router/layouts";
-import { AuthShell, AuthShellFooter } from "./components/auth-shell";
-import { SignUpForm } from "./forms/sign-up-form";
 
 export const signUpRoute = createRoute({
   getParentRoute: () => guestLayout,
   path: "sign-up",
-  component: SignUpPage,
+  component: lazyRouteComponent(() => import("./sign-up.page"), "SignUpPage"),
 });
-
-function SignUpPage() {
-  return (
-    <main>
-      <AuthShell
-        title="Create your account"
-        description="Start building in under a minute."
-        footer={
-          <AuthShellFooter
-            lead="Already have an account?"
-            link={<Link to="/sign-in">Sign in</Link>}
-          />
-        }
-      >
-        <SignUpForm />
-      </AuthShell>
-    </main>
-  );
-}
