+++ apps/app/vite.config.ts
@@ -13,4 +13,5 @@ export default defineConfig({
     include: ["react", "react-dom", "react-dom/client", "react/jsx-runtime"],
   },
   server: { port: 5173 },
+  build: { chunkSizeWarningLimit: 700 },
 });
